--- tests/functional/adapter/incremental/fixtures/expected.py
@@ -1,58 +1,0 @@
-INCREMENTAL_OVERWRITE_TIME_EXPECTED = """
-id,date_hour
-10,2020-01-01 01:00:00
-20,2020-01-01 01:00:00
-30,2020-01-01 02:00:00
-40,2020-01-01 02:00:00
-"""
-
-INCREMENTAL_OVERWRITE_WITH_TIME_EXPECTED = """
-id,date_time
-10,2020-01-01 00:00:00
-20,2020-01-01 00:00:00
-30,2020-01-02 00:00:00
-40,2020-01-02 00:00:00
-"""
-
-INCREMENTAL_OVERWRITE_WITH_DATE_EXPECTED = """
-id,date_day
-10,2020-01-01
-20,2020-01-01
-30,2020-01-02
-40,2020-01-02
-"""
-
-INCREMENTAL_OVERWRITE_WITH_ID_EXPECTED = """
-id
-10
-20
-30
-40
-"""
-
-RANGE_EXPECTED = """
-id,date_int
-10,20200101
-20,20200101
-30,20200102
-40,20200102
-"""
-
-MERGE_EXPECTED = """
-id,date_time
-1,2020-01-01 00:00:00
-2,2020-01-01 00:00:00
-3,2020-01-01 00:00:00
-4,2020-01-02 00:00:00
-5,2020-01-02 00:00:00
-6,2020-01-02 00:00:00
-"""
-
-INCREMENTAL_SEEDS = {
-    'merge_expected.csv': MERGE_EXPECTED,
-    'range_expected.csv': RANGE_EXPECTED,
-    'incremental_overwrite_with_id_expected.csv': INCREMENTAL_OVERWRITE_WITH_ID_EXPECTED,
-    'incremental_overwrite_with_date_expected.csv': INCREMENTAL_OVERWRITE_WITH_DATE_EXPECTED,
-    'incremental_overwrite_with_time_expected.csv': INCREMENTAL_OVERWRITE_WITH_TIME_EXPECTED,
-    'incremental_overwrite_time_expected.csv': INCREMENTAL_OVERWRITE_TIME_EXPECTED
-}
