+++ tests/functional/adapter/incremental/incremental_strategy_fixtures.py
