+++ tests/functional/adapter/simple_bigquery_view/fixtures.py
